+++ R/data-survey.R
@@ -116,6 +116,19 @@ expand_survey_clusters <- function(survey_clusters,
 #' @param age_group_id Age group id.
 #' @param area_top_level Area top level.
 #' @param area_bottom_level Area bottom level.
+#' @param artcov_definition Definition to use for calculate ART coverage.
+#' @param by_restype Whether to stratify estimates by urban/rural restype; logical.
+#'
+#'
+#' @details
+#'
+#' The argument `artcov_definition` controls whether to use both ARV biomarker and
+#' self-report (`artcov_definition = "both"`; default), ARV biomarker only
+#' (`artcov_definition = "arv"`), or self-report ART use only
+#' (`artcov_definition = "artself"`).  If option is `"both"`, then all HIV positive
+#' are used as the denomiator and no missing data on either indicator are
+#' incorporated. If the option is `"arv"` or `"artself"` then missing values in those
+#' variables, respectively, are treated as missing.
 #'
 #' @export
 calc_survey_hiv_indicators <- function(survey_meta,
@@ -127,7 +140,9 @@ calc_survey_hiv_indicators <- function(survey_meta,
                                        sex = c("male", "female", "both"),
                                        age_group_id = NULL,
                                        area_top_level = min(areas$area_level),
-                                       area_bottom_level = max(areas$area_level)) {
+                                       area_bottom_level = max(areas$area_level),
+                                       artcov_definition = c("both", "arv", "artself"),
+                                       by_restype = FALSE) {
 
   ## 1. Identify age groups to calculate for each survey_id
   age_group <- get_age_groups()
@@ -169,7 +184,7 @@ calc_survey_hiv_indicators <- function(survey_meta,
 
   ## 3. Expand individuals dataset to repeat for all individiuals within each
   ##    age/sex group for a given survey
-
+  
   ind <- survey_individuals %>%
     dplyr::inner_join(survey_biomarker,
                       by = c("survey_id", "cluster_id", "household", "line")) %>%
@@ -185,18 +200,34 @@ calc_survey_hiv_indicators <- function(survey_meta,
 
   ind <- dplyr::inner_join(ind, clust_area, by = c("survey_id", "cluster_id"))
 
+  if(by_restype)
+    ind <- dplyr::bind_rows(ind, dplyr::mutate(ind, restype = "all"))
+  else
+    ind <- dplyr::mutate(ind, restype = "all")
+
 
   ## 5. Construct ART coverage indicator as either self-report or ART biomarker
   ##    and gather to long dataset for each biomarker
 
+  if(artcov_definition[1] == "both") {
+    ind <- ind %>%
+      dplyr::group_by(survey_id) %>%
+      dplyr::mutate(has_artcov = any(!is.na(arv) | !is.na(artself)),
+                    artcov = dplyr::case_when(!has_artcov ~ NA_integer_,
+                                              hivstatus == 0 ~ NA_integer_,
+                                              arv == 1 | artself == 1 ~ 1L,
+                                              TRUE ~ 0L)) %>%
+      dplyr::select(-has_artcov, -artself, -arv) %>%
+      dplyr::ungroup()
+  } else if(artcov_definition[1] %in% c("arv", "artself")) {
+    ind <- dplyr::rename(ind, artcov = artcov_definition[1]) %>%
+      dplyr::mutate(artcov = dplyr::if_else(hivstatus == 0,  NA_integer_, as.integer(artcov)))
+  } else {
+    stop(paste("Invalid artcov_definition value:", artcov_definition[1]))
+  }
+    
+    
   ind <- ind %>%
-    dplyr::group_by(survey_id) %>%
-    dplyr::mutate(has_artcov = any(!is.na(arv) | !is.na(artself)),
-                  artcov = dplyr::case_when(!has_artcov ~ NA_integer_,
-                                            hivstatus == 0 ~ NA_integer_,
-                                            arv == 1 | artself == 1 ~ 1L,
-                                            TRUE ~ 0L)) %>%
-    dplyr::select(-has_artcov, -artself, -arv) %>%
     dplyr::rename(prev = hivstatus) %>%
     tidyr::gather(indicator, est, prev, artcov, vls, recent) %>%
     dplyr::filter(!is.na(est))
@@ -208,13 +239,13 @@ calc_survey_hiv_indicators <- function(survey_meta,
     dplyr::filter(!is.na(hivweight), hivweight > 0)
 
   cnt <- dat %>%
-    dplyr::group_by(indicator, survey_id, area_id, sex, age_group) %>%
+    dplyr::group_by(indicator, survey_id, area_id, restype, sex, age_group) %>%
     dplyr::summarise(n_cluster = dplyr::n_distinct(cluster_id),
                      n_obs = dplyr::n()) %>%
     dplyr::ungroup()
 
   datspl <- dat %>%
-    dplyr::mutate(spl = paste(indicator, survey_id, area_level, sex, age_group)) %>%
+    dplyr::mutate(spl = paste(indicator, survey_id, area_level, restype, sex, age_group)) %>%
     split(.$spl)
 
   do_svymean <- function(df) {
@@ -226,7 +257,7 @@ calc_survey_hiv_indicators <- function(survey_meta,
                              weights = ~hivweight)
 
     survey::svyby(~est,
-                  ~ indicator + survey_id + area_id + sex + age_group,
+                  ~ indicator + survey_id + area_id + restype + sex + age_group,
                   des, survey::svymean)
   }
 
@@ -237,10 +268,10 @@ calc_survey_hiv_indicators <- function(survey_meta,
   val <- cnt %>%
     dplyr::full_join(
              dplyr::bind_rows(est_spl),
-             by = c("indicator", "survey_id", "area_id", "sex", "age_group")
+             by = c("indicator", "survey_id", "area_id", "restype", "sex", "age_group")
            ) %>%
     dplyr::left_join(
-             survey_meta %>% select(survey_id, survey_year),
+             survey_meta %>% select(survey_id, survey_mid_calendar_quarter),
              by = c("survey_id")
     ) %>%
     dplyr::left_join(
@@ -249,17 +280,18 @@ calc_survey_hiv_indicators <- function(survey_meta,
              by = c("area_id")
            ) %>%
     dplyr::arrange(
-             fct_relevel(indicator, "prev", "artcov", "vls", "recent"),
+             factor(indicator, c("prev", "artcov", "vls", "recent")),
              survey_id,
-             survey_year,
+             survey_mid_calendar_quarter,
              area_level,
              area_sort_order,
              area_id,
-             fct_relevel(sex, "both", "male", "female"),
+             factor(restype, c("all", "urban", "rural")),
+             factor(sex, c("both", "male", "female")),
              age_group
            ) %>%
     dplyr::select(
-             indicator, survey_id, survey_year, area_id, sex, age_group,
+             indicator, survey_id, survey_mid_calendar_quarter, area_id, restype, sex, age_group,
              n_cluster, n_obs, est, se) %>%
     dplyr::distinct() %>%
     dplyr::mutate(
@@ -268,4 +300,54 @@ calc_survey_hiv_indicators <- function(survey_meta,
            )
 
   val
+}
+
+
+#' Find Calendar Quarter Midpoint of Two Dates
+#'
+#' @param start_date vector coercibel to Date
+#' @param end_date vector coercibel to Date
+#'
+#' @return A vector of calendar quarters
+#'
+#' @examples
+#' start <- c("2005-04-01", "2010-12-01", "2016-01-01")
+#' end <-c("2005-08-01", "2011-05-01", "2016-06-01")
+#'
+#' mid_calendar_quarter <- get_mid_calendar_quarter(start, end)
+#'
+#' @export
+get_mid_calendar_quarter <- function(start_date, end_date) {
+
+  start_date <- lubridate::decimal_date(as.Date(start_date))
+  end_date <- lubridate::decimal_date(as.Date(end_date))
+
+  stopifnot(!is.na(start_date))
+  stopifnot(!is.na(end_date))
+  stopifnot(start_date <= end_date)
+  
+  date4 <- (start_date + end_date) / 2
+  year <- floor(date4) 
+  quarter <- floor((date4 %% 1) * 4) + 1
+
+  paste0("CY", year, "Q", quarter)
+}
+
+
+#' Read Multiple Shape Files in ZIP Archive
+#'
+#' Reads all files in ZIP archive `zfile` matching `pattern` with
+#' function `read_fn` and returns as a list.
+#'
+#' @param zfile path to a zip directory
+#' @param pattern string pattern passed to [`list.files`].
+#' @param read_fn function used to read matched files.
+#'
+#' @export
+read_sf_zip_list <- function(zfile, pattern = "\\.shp$", read_fn = sf::read_sf) {
+  tmpd <- tempfile()
+  on.exit(unlink(tmpd))
+  unzip(zfile, exdir = tmpd)
+  f <- list.files(tmpd, pattern, recursive = TRUE, full.names = TRUE)
+  lapply(f, sf::read_sf)
 }