+++ NEWS.md
@@ -1,3 +1,8 @@
+# naomi 0.0.56
+
+* Replace `survey_year` with `survey_mid_calendar_quarter` in the survey indicators dataset.
+* Add column `restype` for urban/rural to survey indicators datasets and option to stratify survey indicators calculation by urban/rural.
+
 # naomi 0.0.55
 
 * Option for time-varying ART attendance odds ratio. The option `artattend_t2` can be set to TRUE if there is evidence that patterns of cross-district ART attendance changed for the period before the survey and the inter-survey period.